Part of work

+++ src/main/java/org/springframework/samples/petclinic/owner/PetRepository.java
@@ -36,7 +36,7 @@ public interface PetRepository extends Repository<Pet, Integer> {
 
 	/**
 	 * Retrieve all {@link PetType}s from the data store.
-	 * 
+	 *
 	 * @return a Collection of {@link PetType}s.
 	 */
 	@Query("select * from pet_type order by name")
@@ -44,11 +44,11 @@ public interface PetRepository extends Repository<Pet, Integer> {
 	List<PetType> findPetTypes();
 
 	@Query("select * from pet_type where id = :typeId")
-	PetType findPetType(@Param("typeId") Integer id);
+	PetType findPetType(@Param("typeId") Long id);
 
 	/**
 	 * Retrieve a {@link Pet} from the data store by id.
-	 * 
+	 *
 	 * @param id the id to search for
 	 * @return the {@link Pet} if found
 	 */
@@ -57,7 +57,7 @@ public interface PetRepository extends Repository<Pet, Integer> {
 
 	/**
 	 * Save a {@link Pet} to the data store, either inserting or updating it.
-	 * 
+	 *
 	 * @param pet the {@link Pet} to save
 	 */
 	void save(Pet pet);
